requirements.txt
flask
flask-cors
openai
python-dotenv
streamlit
pillow
requests